Add FallbackReplicaOptions
